requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
stripe==7.8.0
googlemaps==4.10.0
django-cors-headers==4.3.1
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.0